+++ api-golf-friends/app/models/user.rb
@@ -1,4 +1,3 @@
 class User < ApplicationRecord
-  has_many :scores, dependent: :delete_all
-
+  has_many :scores
 end
