+++ api-golf-friends/app/serializers/score_serializer.rb
@@ -1,4 +1,4 @@
 class ScoreSerializer < ApplicationSerializer
-  attributes :strokes, :created_at, :user_id, :course_id
+  attributes :strokes, :created_at, :user_id
   has_one :course
 end
