--- api-golf-friends/app/views/api/v1/users/_user.json.jbuilder
@@ -1,3 +1,0 @@
-json.extract! user, :id,
-              :name,
-              :scores
