--- api-golf-friends/app/views/api/v1/users/index.json.jbuilder
@@ -1,5 +1,0 @@
-@users.each do |user|
-  json.set! user.id do
-    json.partial! "api/v1/users/user", user: user
-  end
-end
