+++ api-golf-friends/config/application.rb
@@ -33,8 +33,9 @@ module ApiGolfFriends
     config.middleware.insert_before 0, Rack::Cors do
       allow do
         origins '*'
-        resource '*', headers: :any, methods: [:get, :post, :options, :put, :delete]
+        resource '*', headers: :any, methods: [:get, :post, :options, :put]
       end
     end
+    
   end
 end
