+++ api-golf-friends/config/routes.rb
@@ -1,6 +1,6 @@
 Rails.application.routes.draw do
   namespace :api do
-    namespace :v1, defaults: { format: :json } do
+    namespace :v1 do
       get 'users/:id/scores' => 'users#scores'
       resources :users
       resources :scores
