+++ golffriends-front/src/App.css
@@ -14,22 +14,15 @@
   padding: 20px;
   color: white;
   letter-spacing: 5px;
-    margin-bottom: 2em;
+
 }
 
 .App-title {
-  font-size: 2em;
-  font-family: karla;
+  font-size: 1.5em;
 }
 
 .App-intro {
   font-size: large;
-}
-
-.add {
-  font-size: 30px;
-  background-color: none;
-  color: white;
 }
 
 .modal-background {
@@ -88,12 +81,4 @@ td {
 .modal-button {
   height: 100%;
   width: 100%;
-}
-
-nav > a {
-  font-size: 18px;
-  padding: 20px;
-  letter-spacing: 1px;
-  font-weight: bold;
-  font-family: karla;
 }
