+++ golffriends-front/src/Modal.js
@@ -1,59 +1,114 @@
-import React, { Component } from 'react'
-import Portal from './BPortal';
+import React, { Component } from 'react';
+import ReactDOM from 'react-dom';
 import styled from 'styled-components';
+import {
+  BrowserRouter as Router,
+  Route,
+} from 'react-router-dom';
 
+import ScoreForm from './ScoreForm';
+import RoundHistory from './RoundHistory';
+import ModalToggle from './ModalToggle';
+
+const routes = [
+  {
+    path: '/score-form',
+    component: ScoreForm
+  },
+  {
+    path: '/history',
+    component: RoundHistory
+  },
+]
 
 export default class Modal extends Component {
+  constructor(props) {
+    super(props)
+    this.element = document.createElement('div')
+    this.modalRoot = document.getElementById('modal-root')
+    this.modalRoot.appendChild(this.element)
+  }
+  componentDidMount() {
+    document.addEventListener('keyup', this.keyup, false)
+  }
+  componentWillUnmount() {
+    document.removeEventListener('keyup', this.keyup, false)
+    this.modalRoot.removeChild(this.element)
+  }
+  keyup = e => {
+    if (e.key === 'Escape') {
+      this.props.onClose()
+    }
+  }
+  clickedBackground = () => {
+    this.props.onClose()
+  }
+
+  calculateDiffs = (scores) => {
+    scores.map(score => {
+      return ((score.strokes - score.course.rating)*113 / score.course.slope);
+    });
+  }
+
+  calculateHandicap = () => {
+    if(this.props.scores.length < 5) {
+      return 'N/A';
+    };
+
+    const scores = this.props.scores;
+    const diffs = scores.map(score => {
+      return ((score.strokes - score.course.rating)*113 / score.course.slope);
+    });
+    const sorted = diffs.sort((a, b) => a>b ? 1 : -1 );
+
+    if (sorted.length < 11) {
+      return Math.round(sorted[0]*.96);
+    } else if (sorted.length < 20)  {
+      const lowest = sorted.slice(0, 2);
+      const lowestSum = lowest.reduce((a, b) => a+=b);
+      return Math.round((lowestSum / 3)*.96)
+    } else {
+      const lowest = sorted.slice(0, 9);
+      const lowestSum = lowest.reduce((a, b) => a+=b);
+      return Math.round((lowestSum / 10)*.96)
+    }
+  }
+
   render() {
-      const { children, toggle, on } = this.props;
+    return ReactDOM.createPortal( this._renderModal(), this.element )
+  }
+  _renderModal() {
     return (
-      <Portal>
-        { on &&
-            <ModalWrapper>
-              <ModalCard>
-                <div>{children}</div>
-              </ModalCard>
-              <Background onClick={toggle}/>
-            </ModalWrapper>
-        }
-      </Portal>
+      <div className='modal-background' onClick={this.clickedBackground}>
+        <ModalContainer onClick={e => e.stopPropagation()}>
+          <div className='modal-head'>{this.calculateHandicap()}</div>
+          <Router>
+          <div className='modal-body'>
+              <ModalToggle />
+              {routes.map(({ path, component: C }) => (
+                <Route
+                  key={path}
+                  path={path}
+                  render={(props) => <C {...props} scores={this.props.scores} user={this.props.user}/>}
+                />
+              ))}
+          </div>
+          </Router>
+        </ModalContainer>
+      </div>
     );
   }
 }
 
-const ModalWrapper = styled.div`
-    position: absolute;
-    top: 0;
-    left: 0;
-    height: 100%;
-    width: 100%;
-    display: flex;
-    justify-content: center;
-    align-items:center;
-`;
+const ModalContainer = styled.div`
+  background: white;
+  border-radius: 10px;
+  width: 25vw;
+  min-width:350px;
+  height: 70vh;
 
-const ModalCard = styled.div`
-    position: relative;
-    background: white;
-    border-radius: 5px;
-    box-shadow: 2px 2px 10px rgba(0,0,0, 0.3);
-    z-index:10;
-    min-width: 300px;
-    min-height: 500px;
-    display: flex;
+  color: green;
+  display: flex;
+  flex-direction: column;
 
-    div {
-      width: 100%;
-      
-    }
-`;
-
-const Background = styled.div`
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    top: 0;
-    left 0;
-    background: black;
-    opacity: 0.4;
 `;
