+++ golffriends-front/src/ScoreTile.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React, { Component, Fragment } from 'react';
 import styled from 'styled-components';
 
 export default class ScoreTile extends Component {
@@ -6,22 +6,21 @@ export default class ScoreTile extends Component {
   render() {
     return(
       <TileWrapper>
-        <ScoreDis>{this.props.strokes}</ScoreDis>
-        <CourseDate>{this.props.created_at}<span>{this.props.course}</span></CourseDate>
+        <ScoreDis>88</ScoreDis>
+        <CourseDate>Timberlin White <span>5/11</span></CourseDate>
       </TileWrapper>
     );
   }
 }
 
 const TileWrapper = styled.div`
-  flex: 0 0 auto;
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  margin: 1px;
+  margin: 2px;
   border: 1px solid green;
-  width: 115px;
+  flex-grow: 1;
 `;
 const ScoreDis = styled.div`
   flex-grow:3;
